requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
Django==3.2.6
django-bootstrap4==3.0.1
django-grappelli==2.15.1
django-ranged-response==0.2.0
Pillow==8.3.1
python-decouple==3.4
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
